requirements.txt
Django>=6.0,<7
requests>=2.31
pypdf>=4.0
python-pptx>=0.6.23
